+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 2.9.6
 
+ - Add "x, y, z" and "dx, dy, dz" parameters to all spells, to force source location and direction. (Scripting!)
+ - Add "tx, ty, tz" parameters for target location.
+ - Allow special materials in "material" parameter (e.g. clone, erase, schematics)
+ - Add "Extend" spell, wip.
+ - Add "bypass_pvp", "bypass_build", "cost_reduction" and "cooldown_reduction" parameters.
+
 ## 2.9.5
 
  - Add Factions support for build perms.